feat(nx-generators): add prismerge init generator
