--- libs/nx-generators/src/generators/prismerge/files/src/index.ts__template__
@@ -1,1 +1,0 @@
-const variable = "<%= projectName %>";
