+++ libs/nx-generators/src/generators/prismerge/init/files/__prismergeFileName____template__
@@ -1,0 +1,4 @@
+{
+  "input": [],
+  "output": "<%= prismaSchemaFilePath %>"
+}
