+++ libs/nx-generators/src/generators/prismerge/init/schema.d.ts
@@ -1,0 +1,5 @@
+export interface InitGeneratorSchema {
+  name: string;
+  directory?: string;
+  output: string;
+}
