--- libs/nx-generators/src/generators/prismerge/schema.d.ts
@@ -1,6 +1,0 @@
-export interface NxGeneratorsGeneratorSchema {
-  name: string;
-  library: string;
-  directory: string;
-  prismergeFile: string;
-}
