--- libs/nx-generators/src/generators/prismerge/schema.json
@@ -1,37 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema",
-  "cli": "nx",
-  "$id": "NxGenerators",
-  "title": "",
-  "type": "object",
-  "properties": {
-    "name": {
-      "type": "string",
-      "description": "Name of the Prisma Model to be generated.",
-      "$default": {
-        "$source": "argv",
-        "index": 0
-      }
-    },
-    "library": {
-      "type": "string",
-      "description": "The library to add the prisma file to.",
-      "alias": "lib",
-      "$default": {
-        "$source": "libraryName"
-      },
-      "x-prompt": "What is the name of the library for this model?"
-    },
-    "directory": {
-      "type": "string",
-      "description": "A directory where the model is created.",
-      "alias": "dir",
-      "x-prompt": "In which subdirectory should the model be created?"
-    },
-    "prismergeFile": {
-      "type": "string",
-      "description": "Location of Prismerge File"
-    }
-  },
-  "required": ["name", "library", "directory", "prismergeFile"]
-}
